requirements.txt
google-adk

python-dotenv

requests

# Optional: if you ever parse or pretty-print JSON responses
json5

google-auth
google-auth-httplib2

tqdm
